feat: extract Plug class to plug.py file
